Lab05/lab5_kafka/pyproject.toml
[tool.poetry]
name = "lab5_kafka"
version = "0.1.0"
description = ""
authors = ["Hugo Sousa <sousahugo@ua.pt>"]

[tool.poetry.dependencies]
python = "^3.10"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
